# sci-paper-microservice-silvera/output/SciPaperService/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "SciPaperService", "description": ""}, "servers": [{"url": "http://localhost:9095/api/scipaper"}], "paths": {"/scipaper": {"post": {"description": "Creates a new scipaper.", "operationId": "createSciPaper", "requestBody": {"description": "SciPaper to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}}, "/scipaper/id": {"put": {"description": "Update SciPaper object.", "operationId": "updateSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}, "get": {"description": "Read SciPaper object.", "operationId": "readSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}, "delete": {"description": "Delete SciPaper object.", "operationId": "deleteSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}}, "/createpaper": {"post": {"description": "", "operationId": "createPaper", "parameters": [], "requestBody": {"description": "JSON representation object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "responses": {"200": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SciPaper"}}}}}}}, "/publish": {"post": {"description": "", "operationId": "publish", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "username", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of boolean object", "content": {"application/json": {"schema": {"type": "boolean"}}}}}}}}, "components": {"schemas": {"Section": {"type": "object", "properties": {"name": {"type": "string"}, "content": {"type": "string"}}}, "SciPaper": {"type": "object", "properties": {"author": {"type": "string"}, "title": {"type": "string"}, "sections": {"type": "array", "items": {"$ref": "#/components/schemas/Section"}}}}}, "requestBodies": {"SectionBody": {"description": "JSON representation of Section object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}, "SciPaperBody": {"description": "JSON representation of SciPaper object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}}, "responses": {"SectionOK": {"description": "JSON representation of Section object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}, "SciPaperOK": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}}}}