# sci-paper-microservice-silvera/output/SciPaperService/src/main/java/com/silvera/SciPaperService/domain/model/SciPaper.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-21 14:35:59
*/

package com.silvera.SciPaperService.domain.model;

import org.springframework.data.annotation.Id;
import javax.validation.constraints.*;
public class SciPaper {

    
    @Id
    private String id;
    
    
    private java.lang.String author;
    
    private java.lang.String title;
    
    private java.util.List<Section> sections= java.util.Collections.emptyList();


    
    public java.lang.String getAuthor() {
        return this.author;
    }

    public void setAuthor(java.lang.String author) {
        this.author = author;
    }
    
    public java.lang.String getTitle() {
        return this.title;
    }

    public void setTitle(java.lang.String title) {
        this.title = title;
    }
    
    public java.util.List<Section> getSections() {
        return this.sections;
    }

    public void setSections(java.util.List<Section> sections) {
        this.sections = sections;
    }
    

    
    public String getId(){
        return this.id;
    }
    

}