# sci-paper-microservice-silvera/output/SciPaperService/src/main/java/com/silvera/SciPaperService/domain/model/Section.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-21 14:35:59
*/

package com.silvera.SciPaperService.domain.model;

import org.springframework.data.annotation.Id;
import javax.validation.constraints.*;
public class Section {

    
    @Id
    private String id;
    
    
    private java.lang.String name;
    
    private java.lang.String content;


    
    public java.lang.String getName() {
        return this.name;
    }

    public void setName(java.lang.String name) {
        this.name = name;
    }
    
    public java.lang.String getContent() {
        return this.content;
    }

    public void setContent(java.lang.String content) {
        this.content = content;
    }
    

    
    public String getId(){
        return this.id;
    }
    

}